✨ Add PyFlow support and split package into files

+++ python/virtualenv.go
@@ -1,85 +1,39 @@
 package python
 
 import (
-	"fmt"
 	"os"
-	"os/exec"
 
 	log "github.com/sirupsen/logrus"
 )
 
-// CleanUpVirtualEnv Cleans up a virtual env by deleting .venv/ dir
+// CleanUpVirtualEnv Cleans up a virtual env by deleting .venv/ and __pypackages__/ dirs
 func CleanUpVirtualEnv() {
+	// cleanup __pypackages__ dir
+	if dirExists("./__pypackages__") {
+		log.Debug("Deleting '__pypackages__/' directory")
+
+		os.RemoveAll("./__pypackages__")
+	}
+
 	// cleanup virtual env
-	os.RemoveAll("./.venv")
+	if dirExists("./.venv") {
+		log.Debug("Deleting '.venv/' directory")
+
+		os.RemoveAll("./.venv")
+	}
 }
 
-// SetupVirtualEnv Creates a virtual environment using Poetry, pipenv or pip/venv
+// SetupVirtualEnv Creates a virtual environment using Poetry, PyFlow, pipenv or pip/venv
 func SetupVirtualEnv() {
 	if fileExists("pyproject.toml") {
-		poetryProc()
+		if isPoetry() {
+			poetryProc()
+		} else if isPyFlow() {
+			pyflowProc()
+		}
 	} else if fileExists("Pipfile") {
 		pipenvProc()
 	} else if fileExists("requirements.txt") {
 		pipProc()
 	}
-}
-
-func poetryProc() {
-	log.Info("Found 'pyproject.toml'. Creating virtual environment using 'Poetry'.")
-
-	// Make sure .venv will be in project
-	if !fileExists("poetry.toml") {
-		cmd := exec.Command("poetry", "config", "--local", "virtualenvs.in-project", "true")
-		out, err := cmd.CombinedOutput()
-		if err != nil {
-			log.Error(fmt.Sprintf("%s", err.Error()))
-		}
-		log.Debug(string(out))
-	}
-
-	// install dependencies
-	cmd := exec.Command("poetry", "install")
-	out, err := cmd.CombinedOutput()
-	if err != nil {
-		log.Error(fmt.Sprintf("%s", err.Error()))
-	}
-	log.Debug(string(out))
-}
-
-func pipenvProc() {
-	log.Info("Found 'Pipfile'. Creating virtual environment using 'pipenv'.")
-
-	// Make sure .venv will be in project
-	os.Setenv("PIPENV_VENV_IN_PROJECT", "1")
-
-	// install dependencies
-	cmd := exec.Command("pipenv", "install")
-	out, err := cmd.CombinedOutput()
-	if err != nil {
-		log.Error(fmt.Sprintf("%s", err.Error()))
-	}
-	log.Debug(string(out))
-}
-
-func pipProc() {
-	log.Info("Found 'requirements.txt'. Creating virtual environment using 'pip' & 'venv' module.")
-
-	var cmd *exec.Cmd
-
-	// create virtual env
-	cmd = exec.Command("python3", "-m", "venv", ".venv")
-	out, err := cmd.CombinedOutput()
-	if err != nil {
-		log.Error(fmt.Sprintf("%s", err.Error()))
-	}
-	log.Debug(string(out))
-
-	// install dependencies
-	cmd = exec.Command("./.venv/bin/pip", "install", "-r", "requirements.txt")
-	out, err = cmd.CombinedOutput()
-	if err != nil {
-		log.Error(fmt.Sprintf("%s", err.Error()))
-	}
-	log.Debug(string(out))
 }
